Waiting until the image has stabilized.
This patch introduces a new strategy for capturing the image: the user
is now expected to move the object and then stop moving it. We only
capture the object once the user has stopped moving it. We expect that
this will help capture higher quality images.

Introducing --buffer-stable-frames, --buffer-stability to parameterize
this.xs