Reintroducing object shaking.
Previous commits had removed object shaking as it was too slow for
general use. This patch reintroduces it, as an attempt to improve the
quality of captured images. To make the operation faster, we run the
background extractor on a downsampled image (with default options, we
divide by 3 on both dimentions).

According to my tests, this results in a total cost of ~3 seconds with
default arguments and it manages to remove most of the background. At
this stage, I haven't tested whether removing the background helps
with matching.